backend/requirements.txt
asgiref==3.9.1
blessings==1.7
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.1.7
djangorestframework==3.16.0
django-cors-headers==4.9.0
earthengine-api==1.6.6
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
httplib2==0.31.0
idna==3.10
Pillow==11.3.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.4
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
torch==2.8.0
torchvision==0.23.0
scikit-learn==1.6.1
matplotlib==3.10.0
seaborn==0.13.2
pandas==2.2.3
numpy==2.1.2
